fix: bug

+++ docs/runs/model_codh/detection2.py
@@ -99,29 +99,22 @@ for c in tqdm(range(len(canvases))):
 
     tmp_path = "output/{}/detection/{}.jpg".format(item_id, str(c + 1).zfill(4))
 
-    print("a")
-
-    # 壊れている場合
-    try:
-        print("b")
-        img = Image.open(tmp_path)
-        img.verify()
-    except Exception as e:
-        os.remove(tmp_path)
-        print(e)
-
-    print("c")
+    if os.path.exists(tmp_path):
+        # 壊れている場合
+        try:
+            img = Image.open(tmp_path)
+            img.verify()
+        except Exception as e:
+            os.remove(tmp_path)
+            print(e)
 
     if not os.path.exists(tmp_path):
-        print("d")
-
         os.makedirs(os.path.dirname(tmp_path), exist_ok=True)
 
         # ダウンロードが完了するまで繰り返し
         flg = True
         while flg:
             try:
-                print("e")
                 opener=request.build_opener()
                 opener.addheaders=[('User-Agent','Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/36.0.1941.0 Safari/537.36')]
                 request.install_opener(opener)
@@ -131,7 +124,6 @@ for c in tqdm(range(len(canvases))):
                 os.remove(tmp_path)
                 print(e)
 
-    print("f")
     img = Image.open(tmp_path)
 
     w, h = img.size
